\name{highfrequencyGSOC-package}
\alias{highfrequencyGSOC-package}
\alias{highfrequencyGSOC}
\docType{package}
\title{
Additional functionality for the highfrequency package
}

\description{Extra functionality for the highfrequency package, written by Giang, during the Google Summer of Code.}

\details{
\tabular{ll}{
Package: \tab highfrequencyGSOC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-02\cr
License: \tab GPL\cr
}
}

\author{
Giang Nguyen, Kris Boudt, Jonathan Cornelissen
Maintainer: Giang Nguyen <hagiang3108@gmail.com>
}
